Fix a couple of GCC warnings
operands.c: address_registers and data_registers are not of the same sign, so
returning one of them is awkward.  But, since we always seem to cast the return
value as uint32_t anyway, might as well convert the data_register directly.

renderer.c: pointer-to-int conversion.  The fix is a bit of a mouthful... not
sure it's the canonical way either, but it works.

+++ megado/m68k/operands.c
@@ -366,7 +366,7 @@ Operand* operand_make_address_register_indirect_predec(int n, struct Instruction
 // The data is located at the stored address + a displacement (extension).
 // Extension word format: https://github.com/traviscross/libzrtp/blob/master/third_party/bnlib/lbn68000.c#L342
 
-#define INDEX_REGISTER(extension) ((BIT((extension), 15) ? ctx->address_registers[FRAGMENT((extension), 14, 12)] : ctx->data_registers[FRAGMENT((extension), 14, 12)]))
+#define INDEX_REGISTER(extension) ((BIT((extension), 15) ? ctx->address_registers[FRAGMENT((extension), 14, 12)] : (uint32_t)ctx->data_registers[FRAGMENT((extension), 14, 12)]))
 #define INDEX_LENGTH(extension) (BIT(extension, 11))
 #define INDEX_DISPLACEMENT(extension) (FRAGMENT(extension, 7, 0))
